requirements.txt
streamlit
langchain
openai
python-dotenv
requests
uuid
fpdf
faiss-cpu
sentence-transformers
tiktoken